% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{production}
\alias{production}
\title{Description}
\format{
\tabular{ll}{
   Variable \tab Description \cr
   \code{var} \tab Description \cr
}
}
\source{
Raw data downloaded from \href{url}{text}
}
\usage{
data(production)
}
\description{
Detailed description
}
\examples{
data(production)

head(production)
}
\keyword{datasets}
