% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sales_country}
\alias{sales_country}
\title{Vehicle sales statistics by country and type}
\format{
\tabular{ll}{
   Variable \tab Description \cr
   \code{var} \tab Description \cr
}
}
\source{
Raw data downloaded from the
\href{https://www.oica.net/}{Organisation Internationale des Constructeurs d'Automobiles (OICA)}
and the
\href{https://unstats.un.org/unsd/methodology/m49/overview/}{United Nations Statistics Division}
}
\usage{
data(sales_country)
}
\description{
This data frame contains annual data on vehicle sales statistics by
country and type between the years 2005 and 2021 (inclusive) from the OICA.
Standard country / area codes come from the United Nations Statistics
Division. The table below shows a detailed description of each variable.
}
\examples{
data(sales_country)

head(sales_country)
}
\keyword{datasets}
