% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sales_region}
\alias{sales_region}
\title{Description}
\format{
\tabular{ll}{
   Variable \tab Description \cr
   \code{var} \tab Description \cr
}
}
\source{
Raw data downloaded from \href{url}{text}
}
\usage{
data(sales_region)
}
\description{
Detailed description
}
\examples{
data(sales_region)

head(sales_region)
}
\keyword{datasets}
